# Faster-RCNN from TensorFlow on CPU with Intel's Inference Engine backend