<!-- Please use this line to close one or multiple issues when this pullrequest gets merged
You can add another line right under the first one:
resolves #1234
resolves #1235
-->

### This pullrequest changes

* resolves https://github.com/opencv/opencv/issues/11983

median time for a single 800x600 image forward pass:

| Model | OpenCV on CPU | OpenCV on GPU, fp32 | with IE on CPU |
|---|---|---|---|
| Inception_v2_Faster_RCNN (TensorFlow) **before PR** | 286.62ms | 333.81ms | |
| Inception_v2_Faster_RCNN (TensorFlow) **with PR** | 287.17ms | 335.21ms | 203.22ms |

**Merge with extra**: https://github.com/opencv/opencv_extra/pull/498

```
force_builders=Custom
docker_image:Custom=ubuntu-openvino:16.04
buildworker:Custom=linux-2
test_opencl:Custom=ON
```